move the Config class for the single pulse extractor into the SinglePulseExtractor class itself. This way the Config class is in the 'namespace' of the SinglePulseExtracor class
